pyproject.toml
##########################################################
#              !!! REQUIRES UV >= 0.5.3 !!!              #
##########################################################

[project]
name = "anchored_representation_clouds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "kornia>=0.8.0",
    "lightning>=2.5.0.post0",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "scikit-image>=0.25.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "seaborn>=0.13.2",
    "torch>=2.1.0",
    "torchmetrics>=1.6.1",
    "torchvision>=0.16.0",
    "tqdm>=4.67.1",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cu118" },
]
torchvision = [
  { index = "pytorch-cu118" },
]


[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
